+++ src/affordance/tasks/sequential_interpreter.py
@@ -1,280 +1,11 @@
-import ast
 from dataclasses import dataclass
-import re
-import subprocess
-import sys
 import textwrap
-import time
-from typing import Callable, List, Tuple, Union
+from typing import List, Sequence, Tuple, Union
 
-import enchant
-import openai
 import parsimonious
-from pot_tools import safe_execute, simplify_ans
-import requests
-from serpapi import GoogleSearch
+from tools import Tool, get_tool
 from tqdm import tqdm
-from utils import Command, OpenAIModel, Program, cleanhtml
-
-subscription_key = "28dc412935f24fb9974d80c30915483a"
-search_url = "https://api.bing.microsoft.com/v7.0/search"
-headers = {"Ocp-Apim-Subscription-Key": subscription_key}
-
-
-serp_api_key = "5599fad2263e4b1c6d4efb62fa15f83f5ad2d1bb727b8721d616d719399a7f7b"
-
-
-def search(query, top_k=1):
-    params = {"q": query, "textDecorations": True, "textFormat": "HTML"}
-    response = requests.get(search_url, headers=headers, params=params)
-    response.raise_for_status()
-    search_results = response.json()
-    if "webPages" in search_results:
-        snippets = [
-            cleanhtml(v["name"] + " " + v["snippet"]) for v in search_results["webPages"]["value"]
-        ][:top_k]
-    else:
-        return "", None
-    return "\n".join(snippets), None
-
-
-def google_search(query, previous_input=None, top_k=1):
-    params = {
-        "q": query,
-        #   "location": "Austin, Texas, United States",
-        "hl": "en",
-        "gl": "us",
-        "google_domain": "google.com",
-        "api_key": serp_api_key,
-    }
-
-    search = GoogleSearch(params)
-    try:
-        res = search.get_dict()
-    except:
-        res = {}
-
-    if "answer_box" in res.keys() and "answer" in res["answer_box"].keys():
-        toret = res["answer_box"]["answer"]
-    elif "answer_box" in res.keys() and "snippet" in res["answer_box"].keys():
-        toret = res["answer_box"]["snippet"]
-    elif "answer_box" in res.keys() and "snippet_highlighted_words" in res["answer_box"].keys():
-        toret = res["answer_box"]["snippet_highlighted_words"][0]
-    elif "organic_results" in res.keys() and "snippet" in res["organic_results"][0].keys():
-        toret = res["organic_results"][0]["snippet"]
-    else:
-        toret = None
-
-    return toret, None
-
-
-def code_generate(instruction, code_input):
-    # response = openai.Edit.create(
-    # model="code-davinci-edit-001",
-    # input="x = " + code_input,
-    # instruction="Python code for " + instruction,
-    # temperature=0,
-    # top_p=1
-    # )
-    if instruction in code_input:
-        # Code input is sufficient to write code
-        comment = '"""\n{0}\nStore the final result as a variable named \'ans\' and print it.\n"""\n\n'.format(
-            code_input
-        )
-    else:
-        if re.search("#[0-9]+", instruction):
-            input_string = re.search("#[0-9]+", instruction).group(0)
-            instruction = instruction.replace(input_string, "Input")
-        # comment = "\"\"\"\n{0}\nInput:{1}\nStore the final result as a variable named 'ans' and print it.\n\"\"\"\n\ndef".format(instruction, code_input)
-        comment = '"""\n{0}\nInput:{1}\nStore the final result as a variable named \'ans\' and print it.\n"""\n\n'.format(
-            instruction, code_input
-        )
-
-    response = openai.Completion.create(
-        # model="code-davinci-002",
-        model="davinci-codex-002-msft",
-        prompt=comment,
-        temperature=0.0,
-        max_tokens=256,
-        top_p=1,
-        frequency_penalty=0,
-        presence_penalty=0,
-        stop=["print(ans)"],
-    )
-
-    code_snippet = response.choices[0].text
-    time.sleep(5)
-
-    # TODO: Add function run or remove the "def" entirely
-    # return "def" + code_snippet, comment
-    # if "ans" in code_snippet:
-    #     code_snippet += "\nprint(ans)"
-
-    # TODO: Add "print(ans) to the end."
-    return code_snippet, comment
-
-
-def code_execute(code_snippet, code_input=None):
-    if code_input:
-        try:
-            # Run arithmetic python code generate
-            code_output = safe_execute(code_input)
-            result = simplify_ans(code_output)
-            if result:
-                return result, code_input
-        except:
-            pass
-
-    try:
-        # Run code snippet as is, if it was generated by code_generate
-        # More robust outputs
-        result = subprocess.run(
-            [sys.executable, "-c", code_input],
-            capture_output=True,
-            text=True,
-            timeout=60,
-        )
-        if result.stderr == "":
-            simplified_result = simplify_ans(result.stdout)
-            return simplified_result, code_snippet
-    except:
-        pass
-
-    try:
-        # Run arithmetic python code generate
-        code_output = safe_execute(code_snippet)
-        result = simplify_ans(code_output)
-        if result:
-            return result, code_snippet
-    except:
-        pass
-
-    try:
-        # Run code snippet as is, if it was generated by code_generate
-        result = subprocess.run(
-            [sys.executable, "-c", code_snippet],
-            capture_output=True,
-            text=True,
-            timeout=60,
-        )
-        if result.stderr == "":
-            simplified_result = simplify_ans(result.stdout)
-            return simplified_result, code_snippet
-    except:
-        pass
-
-    if "error" in code_snippet or (code_input and "error" in code_input):
-        return code_error(code_snippet, code_input)
-
-    try:
-        # Command instruction (if its code) starts in the next line.
-        code_snippet = code_snippet.split("\n", 1)[1]
-        result = subprocess.run(
-            [sys.executable, "-c", code_snippet],
-            capture_output=True,
-            text=True,
-            timeout=60,
-        )
-        if result.stderr == "":
-            simplified_result = simplify_ans(result.stdout)
-            return simplified_result, code_snippet
-    except:
-        return None, code_snippet
-
-    return None, code_snippet
-
-
-def code_generate_then_execute(instruction, code_input):
-    # output = openai.Edit.create(
-    # model="code-davinci-edit-001",
-    # input="x = " + code_input,
-    # instruction="Python code for " + instruction,
-    # temperature=0,
-    # top_p=1
-    # )
-    # code_snippet = output.choices[0].text
-
-    code_snippet, generate_details = code_generate(instruction=instruction, code_input=code_input)
-    time.sleep(2)
-    # result = subprocess.run([sys.executable, "-c", code_snippet], capture_output=True, text=True)
-    # if result.stderr == "":
-    #     return result.stdout, code_snippet
-    result = code_execute(code_snippet=code_snippet)
-    return result
-
-
-def code_error(code_snippet, code_input):
-    # Command instruction (if its code) starts in the next line.
-    try:
-        code_snippet = code_snippet.split("\n", 1)[1]
-        result = subprocess.run(
-            [sys.executable, "-c", code_snippet],
-            capture_output=True,
-            text=True,
-            timeout=60,
-        )
-        if result.stderr != "":
-            return result.stderr, code_snippet
-    except:
-        return None, code_snippet
-
-
-def code_edit(instruction, code_input):
-    comment = "# " + instruction + "\n"
-    response = openai.Completion.create(
-        # model="code-davinci-002",
-        model="davinci-codex-002-msft",
-        prompt=comment,
-        temperature=0.0,
-        max_tokens=256,
-        top_p=1,
-        frequency_penalty=0,
-        presence_penalty=0,
-        stop=["\n"],
-    )
-
-    code_snippet = response.choices[0].text
-    return comment + code_snippet, comment
-
-
-def code_generate_then_lookup(instruction, code_input):
-    code_snippet, generate_details = code_generate(instruction=instruction, code_input=code_input)
-    time.sleep(2)
-    result, execute_snippet = code_execute(code_snippet=code_snippet)
-    d = enchant.Dict("en_US")
-    try:
-        word_list = ["".join(chat_list) for chat_list in ast.literal_eval(result)]
-        valid_list = []
-        for word in word_list:
-            if d.check(word):
-                valid_list.append(word)
-        return simplify_ans(valid_list), execute_snippet
-    except:
-        return None, execute_snippet
-
-
-def lookup(word_list):
-    d = enchant.Dict("en_US")
-    valid_list = []
-    for word in word_list:
-        if d.check(word):
-            valid_list.append(word)
-    return valid_list
-
-
-def arithmetic(equations, previous_input):
-    # collect all outputs of arithmetic and run python to execute them
-    result = subprocess.run(
-        [sys.executable, "-c", "\n".join(equations)], capture_output=True, text=True
-    )
-    return result, None
-
-
-_gpt3 = OpenAIModel(model="text-davinci-002", max_length=200, quote="", n=1)
-
-
-def generate(prompt, previous_input):
-    return _gpt3(prompt)[0], None
+from utils import Command, OpenAIModel, Program
 
 
 def trim_demonstrations(s: str) -> str:
@@ -298,31 +29,11 @@ class TopDownVisitor:
         model_name: str = "text-davinci-002",
         temperature: float = 0.3,
         rerun: bool = True,
-        exclude_list: List[str] = [],
+        exclude: Sequence[str] = [],
     ):
-        self.built_ins = {
-            "[generate]": generate,
-            "[search]": google_search,
-            "[code generate]": code_generate,
-            "[code execute]": code_execute,
-            "[code edit]": code_edit,
-            "[string edit]": code_generate_then_execute,
-            "[string index]": code_generate_then_execute,
-            "[string permutation]": code_generate_then_lookup,
-            "[arithmetic]": arithmetic,
-            "[generate python code]": code_generate,
-        }
-        for key in exclude_list:
-            del self.built_ins[key]
-        self.keyword_map = {
-            "[permutation]": "[string permutation]",
-            "[permute]": "[string permutation]",
-            "[string permute]": "[string permutation]",
-            "[execute]": "[code execute]",
-        }
-
-        self.execution_details: List[StacktraceItem] = []
         self.rerun = rerun
+        self.exclude = exclude
+        self.execution_details: List[StacktraceItem] = []
         self.program_completer = OpenAIModel(
             model_name, quote="---", temperature=temperature, max_length=500, n=1
         )
@@ -337,17 +48,6 @@ class TopDownVisitor:
         for prefix, program in tqdm(zip(prefixes, programs)):
             answers.append(self.visit(prefix, program))
         return answers
-
-    def get_builtin(self, command: str) -> Union[Callable[[str, str], Tuple[str, str]], None]:
-        """Get the tool function from its name (or an alias in self.keyword_map).
-
-        Returns None if not found."""
-        if command in self.built_ins:
-            return self.built_ins[command]
-        if command in self.keyword_map:
-            return self.built_ins[self.keyword_map[command]]
-
-        return None
 
     def shorten_prefix(self, prefix: str, to_remove: int) -> str:
         """Remove few-shot examples from the prefix, starting with the latter ones.
@@ -445,7 +145,7 @@ class TopDownVisitor:
                 break
 
             # get the tool function
-            tool_func = self.get_builtin(command.name)
+            tool_func = get_tool(command.name, self.exclude)
 
             if tool_func is None:
                 print(f"tool '{command.name}' not found")
